+++ piv_card.py
@@ -23,6 +23,7 @@ try:
     from smartcard.CardRequest import CardRequest
     from smartcard.util import toBytes, toHexString
     from smartcard.Exceptions import CardRequestTimeoutException
+    from smartcard.pcsc.PCSCExceptions import EstablishContextException
     from smartcard.System import readers
 except ModuleNotFoundError as exc:
     raise ModuleNotFoundError("pyscard not installed or was not found") from exc
@@ -181,12 +182,19 @@ class PIVcard:
     def __init__(self, connect_timeout, debug=False):
         self.debug = debug
         piv_card_atr = ATRCardType(PIVcard.YUBICO5_ATR_HEX)
-        readers()
-        cardrequest = CardRequest(timeout=connect_timeout, cardType=piv_card_atr)
         try:
+            readers()
+            cardrequest = CardRequest(timeout=connect_timeout, cardType=piv_card_atr)
             self.cardservice = cardrequest.waitforcard()
         except CardRequestTimeoutException:
             raise PIVCardTimeoutException
+        except EstablishContextException as exc:
+            if (
+                str(exc)
+                == "'Failure to establish context: The Smart Card Resource Manager is not running. '"
+            ):
+                raise ConnectionException("Can't start Scard service")
+            raise
         self.cardservice.connection.connect()
         apdu_select = [
             0x00,